requirements.txt
Django==2.2.16
django-heroku==0.3.1
gunicorn==20.0.4
python-telegram-bot==13.1
requests==2.22.0